requirements.txt

flask
python-telegram-bot==20.0
mnemonic
eth-account
